Move services reducer to userSelection
