+++ src/client/src/store/combineReducers.ts
@@ -3,7 +3,6 @@ import navigation from "./navigation/combineReducers";
 import templates from "./templates/combineReducers";
 import config from "./config/combineReducers";
 import userSelection from "./userSelection/combineReducers";
-import services from "./azureProfileData";
 import RootAction from "./ActionType";
 import { CONFIG_TYPEKEYS } from "./config/configTypeKeys";
 
@@ -11,7 +10,6 @@ const appReducer = combineReducers({
   templates,
   config,
   userSelection,
-  services,
   navigation
 });
