+++ src/client/src/store/userSelection/combineReducers.ts
@@ -5,13 +5,15 @@ import frontendFramework from "./frameworks/selectFrontendFrameworkReducer";
 import pages from "./pages/selectPagesReducer";
 import projectNameObject from "./app/updateProjectNameReducer";
 import outputPathObject from "./app/updateOutputPathReducer";
+import services from "./services/combineReducers";
 
 const UserSelectionStateReducer = combineReducers({
   frontendFramework,
   backendFramework,
   pages,
   outputPathObject,
-  projectNameObject
+  projectNameObject,
+  services
 });
 
 export default UserSelectionStateReducer;
